Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@9694 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/graphics/particle_emitter.cpp
@@ -336,8 +336,8 @@ void ParticleEmitter::setParticleType(const ParticleKind* type)
         m_node->setName(debug_name.c_str());
     }
 #endif
-    m_min_rate = type->getMinRate();
-    m_max_rate = type->getMaxRate();
+    m_min_rate = (float)type->getMinRate();
+    m_max_rate = (float)type->getMaxRate();
     
     if (isNewType)
     {
